@@ -21,59 +56,110 @@
   },
   "homepage": "https://github.com/createthis/hapijs-react-redux#readme",
   "dependencies": {
-    "axios": "^0.16.2",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-transform-class-properties": "^6.16.0",
+    "@babel/core": "^7.3.3",
+    "@babel/node": "^7.2.2",
+    "@babel/plugin-proposal-private-methods": "^7.4.4",
+    "@babel/polyfill": "^7.2.5",
+    "@babel/preset-env": "=7.4.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
+    "@react-bootstrap/pagination": "^1.0.0",
+    "axios": "^0.19.0",
+    "babel-loader": "^8.0.5",
+    "babel-plugin-add-module-exports": "^1.0.0",
+    "babel-plugin-root-import": "^5.1.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-register": "^6.26.0",
-    "boom": "^5.2.0",
+    "boom": "^7.2.2",
     "bootstrap": "^3.3.7",
-    "chart.js": "^2.3.0",
-    "css-loader": "^0.28.7",
-    "csv": "^1.1.0",
-    "csv-parse": "^1.1.7",
-    "eslint": "^4.7.2",
-    "eslint-plugin-react": "^7.3.0",
-    "handlebars": "^4.0.5",
-    "hapi": "^16.6.0",
-    "hapi-webpack-plugin": "^2.0.0",
-    "hoek": "^4.1.0",
-    "html-webpack-plugin": "^2.22.0",
-    "inert": "^4.0.2",
-    "joi": "^11.0.3",
-    "jquery": "^3.1.1",
-    "minimatch": "^3.0.2",
-    "moment": "^2.15.1",
-    "react": "^15.3.2",
-    "react-bootstrap": "^0.31.3",
-    "react-chartjs-2": "^2.6.2",
-    "react-dom": "^15.3.2",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "react-redux": "^5.0.6",
-    "react-select": "^1.0.0-rc.2",
-    "redux": "^3.6.0",
-    "sequelize": "^4.10.0",
-    "sequelize-cli": "^3.0.0-3",
-    "sqlite3": "^3.1.6",
-    "style-loader": "^0.18.2",
-    "vision": "^4.1.0",
-    "webpack": "^3.6.0"
+    "create-react-class": "^15.6.3",
+    "css-loader": "^2.1.1",
+    "dotenv": "^5.0.1",
+    "form-data": "^2.3.2",
+    "formdata-polyfill": "^3.0.10",
+    "get-port": "^4.1.0",
+    "good": "^8.1.1",
+    "good-console": "github:createthis/good-console",
+    "good-squeeze": "^5.1.0",
+    "handlebars": "4.1.2",
+    "hapi": "^17.7.0",
+    "hapi-auth-basic": "^5.0.0",
+    "hapi-auth-bearer-token": "^6.1.0",
+    "hapi-auth-cookie": "^8.1.0",
+    "hapi-webpack-plugin": "github:createthis/hapi-webpack-plugin",
+    "history": "^4.7.2",
+    "hoek": "^6.0.3",
+    "html-webpack-plugin": "^3.2.0",
+    "inert": "^5.1.2",
+    "istanbul": "^0.4.5",
+    "joi": "^12.0.0",
+    "jquery": "^3.4.0",
+    "mime-types": "^2.1.22",
+    "moment": "^2.22.1",
+    "moment-range": "^4.0.2",
+    "mysql": "^2.15.0",
+    "mysql2": "^1.5.3",
+    "node-rsa": "^1.0.1",
+    "nodemailer": "^5.1.1",
+    "prop-types": "^15.6.2",
+    "qs": "^6.5.2",
+    "rand-token": "^0.4.0",
+    "react": "^16.8.6",
+    "react-bootstrap": "^0.32.1",
+    "react-dom": "^16.8.2",
+    "react-dom-factories": "^1.0.2",
+    "react-hot-loader": "^4.7.1",
+    "react-redux": "^7.0.1",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
+    "redis": "^2.8.0",
+    "redis-server": "^1.2.2",
+    "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0",
+    "sequelize": "^5.12.2",
+    "sequelize-cli": "^5.5.0",
+    "sequelize-paper-trail": "git+https://github.com/libertas-funding/sequelize-paper-trail.git#dist",
+    "sinon": "^5.0.7",
+    "sqlite3": "^4.0.3",
+    "style-loader": "^0.21.0",
+    "vision": "^5.4.3",
+    "webpack": "^4.29.5",
+    "yup": "^0.24.1"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "factory-girl": "^4.0.0-beta.7",
-    "mocha": "^3.1.2",
-    "nodemon": "^1.11.0",
-    "redux-devtools": "^3.3.1",
-    "sequelize-auto": "^0.4.16",
-    "should": "^13.1.0",
-    "webpack-dev-middleware": "^1.12.0"
+    "@babel/cli": "^7.2.3",
+    "@babel/plugin-proposal-class-properties": "^7.3.3",
+    "@babel/plugin-proposal-object-rest-spread": "^7.3.2",
+    "axios-mock-adapter": "^1.15.0",
+    "babel-plugin-css-modules-transform": "^1.6.1",
+    "babel-plugin-istanbul": "^5.1.1",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "enzyme": "^3.9.0",
+    "enzyme-adapter-react-16": "^1.6.0",
+    "eslint": "^5.16.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-react": "^7.11.1",
+    "factory-girl": "^5.0.2",
+    "jsdom": "^11.10.0",
+    "jsdom-global": "^3.0.2",
+    "mocha": "^5.2.0",
+    "mock-sftp-server": "^0.1.0",
+    "nodemailer-mock-transport": "^1.3.0",
+    "nodemon": "^1.18.7",
+    "nyc": "^14.0.0",
+    "react-test-renderer": "^16.8.2",
+    "redux-devtools": "^3.5.0",
+    "redux-mock-store": "^1.5.3",
+    "sequelize-auto": "github:createthis/sequelize-auto",
+    "should": "^13.2.1",
+    "should-enzyme": "2.1.0",
+    "should-sinon": "0.0.6",
+    "timekeeper": "^2.1.2",
+    "webpack-dev-middleware": "^3.6.0"
   },
   "engines": {
-    "node": ">=6.8.0"
+    "node": ">=10.14.1"
   }
 }
