@@ -1,16 +1,16 @@
-import "babel-polyfill";
+import '@babel/polyfill';
 import React from 'react';
 import { render } from 'react-dom'
 import { createStore } from 'redux'
 import { Provider } from 'react-redux'
-import Home from './components/home.jsx';
-import reducer from './reducers'
+import Root from '~/client/components/root.jsx';
+import reducer from '~/client/reducers'
 
 const store = createStore(reducer)
 
 render(
   <Provider store={store}>
-    <Home />
+    <Root />
   </Provider>,
-  document.getElementById('app')
+  document.getElementById('root')
 );
