@@ -1,4 +1,4 @@
-import * as actions from '../../../client/actions/index'
+import * as actions from '~/client/actions/index'
 
 describe('actions', () => {
 })
